Store syntax aliases in serialized scoper infotable
