+++ tests/Compilation/positive/test073/Package.juvix
@@ -1,0 +1,5 @@
+module Package;
+
+import PackageDescription.V2 open;
+
+package : Package := defaultPackage {name := "test073"};
