+++ tests/Compilation/positive/test073/test073.juvix
@@ -1,0 +1,5 @@
+module test073;
+
+import SyntaxAlias open;
+
+main : MyNat := 11;
